refactor(api_summary): isolate JS interop traits into dedicated src/js_trait.dart library
